# openapi/open_api.yml
openapi: 3.0.1
info:
  title: Pagopa payment gateway Mock
  version: '1.0.0'
...
paths:
...
  /xpay/change/autenticazione3DS/outcome:
    post:
      tags:
        - XPay Change Outcome Controller
      operationId: xpayChangeAuthenticationOutcome
      parameters:
        - in: query
          name: outcome
          example: "OK"
          schema:
            type: string
            enum:
              - "OK"
              - "KO"
        - in: query
          name: errorCode
          example: 97L
          schema:
            type: number
      responses:
        "200":
          description: OK
...
  /xpay/change/paga3DS/outcome:
    post:
      tags:
        - XPay Change Outcome Controller
      operationId: xpayChangePaymentOutcome
      parameters:
        - in: query
          name: outcome
          example: "OK"
          schema:
            type: string
            enum:
              - "OK"
              - "KO"
        - in: query
          name: errorCode
          example: 97L
          schema:
            type: number
      responses:
        "200":
          description: OK
...
  /xpay/change/storna/outcome:
    post:
      tags:
        - XPay Change Outcome Controller
      operationId: xPayRefundChangeOutcome
      parameters:
        - name: outcome
          in: query
          required: false
          schema:
            type: string
            enum:
              - OK
              - KO
        - name: errorCode
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
  /xpay/change/situazioneOrdine/outcome:
    post:
      tags:
        - XPay Change Outcome Controller
      operationId: xPayOrderStatusChangeOutcome
      parameters:
        - name: outcome
          in: query
          required: false
          schema:
            type: string
            enum:
              - OK
              - KO
        - name: errorCode
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: string
...